+++ src/ReconNess.Web/ClientApp/src/store/modules/accounts.js
@@ -28,7 +28,21 @@ const actions = {
                 reject(err)
             }
         })
-    }
+    },
+    latestVersion() {
+        return new Promise((resolve, reject) => {
+            try {
+                api.get('accounts/latestVersion')
+                    .then((res) => {
+                        resolve(res.data)
+                    })
+                    .catch(err => reject(err))
+            }
+            catch (err) {
+                reject(err)
+            }
+        })
+    },
 }