+++ src/ReconNess.Web/Controllers/AccountsController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
+using ReconNess.Core.Providers;
 using ReconNess.Core.Services;
 using ReconNess.Entities;
 using ReconNess.Web.Dtos;
@@ -14,14 +15,17 @@ namespace ReconNess.Web.Controllers
     {
         private readonly IMapper mapper;
         private readonly INotificationService notificationService;
+        private readonly IVersionProvider currentVersionProvider;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="AccountsController" /> class
         public AccountsController(IMapper mapper,
-            INotificationService notificationService)
+            INotificationService notificationService,
+            IVersionProvider currentVersionProvider)
         {
             this.mapper = mapper;
             this.notificationService = notificationService;
+            this.currentVersionProvider = currentVersionProvider;
         }
 
         // GET api/account/notification
@@ -44,6 +48,15 @@ namespace ReconNess.Web.Controllers
             await this.notificationService.SaveNotificationAsync(notification, cancellationToken);
 
             return NoContent();
+        }
+
+        // GET api/account/latestVersion
+        [HttpGet("latestVersion")]
+        public async Task<IActionResult> LatestVersion(CancellationToken cancellationToken)
+        {
+            var currentVersion = await this.currentVersionProvider.GetLatestVersionAsync(cancellationToken);
+
+            return Ok(currentVersion);
         }
     }
 }